test: extract provider sections to tests/providers/ (batch D, #1440)
