+++ test-all.mjs
@@ -4996,108 +4996,6 @@ try {
   fail(`custom instructions test crashed: ${e.message}`);
 }
 
-// ── 43. PROVIDERS — NoFluffJobs ───────────────────────────────────────────
-
-console.log('\n43. Provider — nofluffjobs');
-
-try {
-  const nfj = (await import(pathToFileURL(join(ROOT, 'providers/nofluffjobs.mjs')).href)).default;
-  const { parseNoFluffJobsResponse } = await import(pathToFileURL(join(ROOT, 'providers/nofluffjobs.mjs')).href);
-
-  if (nfj.id === 'nofluffjobs') pass('nofluffjobs.id is "nofluffjobs"');
-  else fail(`nofluffjobs.id is ${JSON.stringify(nfj.id)}`);
-
-  if (nfj.detect({ name: 'NoFluff', careers_url: 'https://nofluffjobs.com/pl' })?.url) {
-    pass('nofluffjobs.detect() matches nofluffjobs.com URL');
-  } else {
-    fail('nofluffjobs.detect() should match nofluffjobs.com URL');
-  }
-
-  if (nfj.detect({ name: 'X', careers_url: 'https://evil.example/nofluffjobs.com/pl' }) === null) {
-    pass('nofluffjobs.detect() rejects path-spoofed URLs');
-  } else {
-    fail('nofluffjobs.detect() must reject path-spoofed URLs');
-  }
-
-  const fakeResponse = {
-    postings: [
-      {
-        title: 'Frontend Engineer',
-        name: 'ExampleCo',
-        url: 'frontend-engineer-remote',
-        posted: 1781270000000,
-        fullyRemote: true,
-        location: { places: [{ city: 'Kraków' }] },
-      },
-      { title: '', name: 'Broken', url: 'missing-title' },
-      7,
-    ],
-    totalPages: 1,
-  };
-
-  const parsed = parseNoFluffJobsResponse(fakeResponse);
-  if (parsed.length === 1) pass('nofluffjobs parser filters malformed rows');
-  else fail(`nofluffjobs parser returned ${parsed.length} rows, expected 1`);
-
-  if (parsed[0].title === 'Frontend Engineer' && parsed[0].url === 'https://nofluffjobs.com/pl/job/frontend-engineer-remote') {
-    pass('nofluffjobs parser maps title and URL');
-  } else {
-    fail(`nofluffjobs parser mapped title/url incorrectly: ${JSON.stringify(parsed[0])}`);
-  }
-
-  if (parsed[0].company === 'ExampleCo' && parsed[0].location === 'Remote, Kraków' && parsed[0].postedAt === 1781270000000) {
-    pass('nofluffjobs parser maps company, location, and postedAt');
-  } else {
-    fail(`nofluffjobs parser mapped fields incorrectly: ${JSON.stringify(parsed[0])}`);
-  }
-
-  let capturedUrl = '';
-  let capturedOpts = null;
-  const fetched = await nfj.fetch(
-    { name: 'NoFluffJobs', careers_url: 'https://nofluffjobs.com/pl', max_pages: 1 },
-    {
-      transport: 'http',
-      fetchJson: async (url, opts) => {
-        capturedUrl = url;
-        capturedOpts = opts;
-        return fakeResponse;
-      },
-      fetchText: async () => '',
-    },
-  );
-  if (fetched.length === 1 && capturedUrl.startsWith('https://nofluffjobs.com/api/search/posting?')) {
-    pass('nofluffjobs.fetch() uses search posting API endpoint');
-  } else {
-    fail(`nofluffjobs.fetch() endpoint/result wrong: ${capturedUrl} ${JSON.stringify(fetched)}`);
-  }
-
-  if (capturedOpts && capturedOpts.method === 'POST' && capturedOpts.redirect === 'error') {
-    pass('nofluffjobs.fetch() uses POST and redirect:"error"');
-  } else {
-    fail(`nofluffjobs.fetch() should use POST and redirect:"error", got ${JSON.stringify(capturedOpts)}`);
-  }
-
-  let ssrfRejected = false;
-  try {
-    await nfj.fetch({ name: 'Evil', careers_url: 'https://evil.example/pl' }, { transport: 'http', fetchJson: async () => fakeResponse, fetchText: async () => '' });
-  } catch (e) {
-    if (e.message.includes('trusted nofluffjobs.com')) ssrfRejected = true;
-  }
-  if (ssrfRejected) pass('nofluffjobs.fetch() rejects untrusted host');
-  else fail('nofluffjobs.fetch() should reject untrusted host');
-
-  let badShape = false;
-  try {
-    parseNoFluffJobsResponse({ jobs: [] });
-  } catch (e) {
-    if (e.message.includes('unexpected API response')) badShape = true;
-  }
-  if (badShape) pass('nofluffjobs parser throws on bad response shape');
-  else fail('nofluffjobs parser should throw on bad response shape');
-} catch (e) {
-  fail(`nofluffjobs provider tests crashed: ${e.message}`);
-}
-
 // ── 44. openrouter-runner — portals drift guard ─────────────────
 console.log('\n44. openrouter-runner — portals drift guard');
 
@@ -5205,469 +5103,6 @@ try {
   fail(`cooldown filter tests crashed: ${e.message}`);
 }
 
-// ── 46. Provider — jobspresso ──────────────────────────────────────
-
-console.log('\nXX. Provider — jobspresso');
-
-try {
-  const {
-    default: jobspresso,
-    parseJobspressoFeed,
-  } = await import(pathToFileURL(join(ROOT, 'providers/jobspresso.mjs')).href);
-
-  if (jobspresso.id === 'jobspresso') {
-    pass('jobspresso.id is "jobspresso"');
-  } else {
-    fail(`jobspresso.id is "${jobspresso.id}"`);
-  }
-
-  if (
-    jobspresso.detect({ provider: 'jobspresso' })?.url ===
-      'https://jobspresso.co/?feed=job_feed'
-  ) {
-    pass('jobspresso.detect() claims explicit provider config');
-  } else {
-    fail('jobspresso.detect() failed');
-  }
-
-  if (jobspresso.detect({ provider: 'other' }) === null) {
-    pass('jobspresso.detect() ignores other provider ids');
-  } else {
-    fail('jobspresso.detect() should ignore other providers');
-  }
-
-  const xml = `
-<rss>
-  <channel>
-    <item>
-      <title><![CDATA[Senior Backend Engineer]]></title>
-      <link>https://jobspresso.co/job/acme/backend</link>
-      <pubDate>Mon, 02 Jun 2025 12:00:00 GMT</pubDate>
-      <job_listing:company><![CDATA[Acme]]></job_listing:company>
-      <job_listing:location><![CDATA[Remote]]></job_listing:location>
-    </item>
-
-    <item>
-      <title></title>
-      <link>https://jobspresso.co/job/skip</link>
-    </item>
-
-    <item>
-      <title>Bad Host</title>
-      <link>https://evil.com/job</link>
-    </item>
-  </channel>
-</rss>
-`;
-
-  const jobs = parseJobspressoFeed(xml);
-
-  if (jobs.length === 1) {
-    pass('parseJobspressoFeed keeps valid items and drops malformed ones');
-  } else {
-    fail(`expected 1 job, got ${jobs.length}`);
-  }
-
-  const job = jobs[0];
-
-  if (
-    job.title === 'Senior Backend Engineer' &&
-    job.company === 'Acme' &&
-    job.location === 'Remote' &&
-    job.url === 'https://jobspresso.co/job/acme/backend' &&
-    typeof job.postedAt === 'number'
-  ) {
-    pass('parseJobspressoFeed maps title, url, company, location and postedAt');
-  } else {
-    fail(`unexpected parsed job: ${JSON.stringify(job)}`);
-  }
-
-  let fetchCalled = false;
-
-  const fetched = await jobspresso.fetch({}, {
-    async fetchText(url, opts) {
-      fetchCalled = true;
-
-      if (
-        url !== 'https://jobspresso.co/?feed=job_feed' ||
-        opts?.redirect !== 'error'
-      ) {
-        throw new Error('unexpected fetch arguments');
-      }
-
-      return xml;
-    },
-  });
-
-  if (fetchCalled && fetched.length === 1) {
-    pass('jobspresso.fetch() requests the pinned RSS feed');
-  } else {
-    fail('jobspresso.fetch() did not fetch correctly');
-  }
-
-  if (fetchCalled) {
-    pass('jobspresso.fetch() passes redirect:"error" to fetchText');
-  } else {
-    fail('jobspresso.fetch() never called fetchText');
-  }
-
-} catch (e) {
-  fail(`jobspresso provider tests crashed: ${e.message}`);
-}
-
-// ── 47. Provider — nodesk ───────────────────────────────────────
-console.log('\n47. Provider — nodesk');
-
-try {
-  const {
-    default: nodesk,
-    parseNodeskFeed,
-  } = await import(pathToFileURL(join(ROOT, 'providers/nodesk.mjs')).href);
-
-  if (nodesk.id === 'nodesk') {
-    pass('nodesk.id is "nodesk"');
-  } else {
-    fail(`nodesk.id is "${nodesk.id}"`);
-  }
-
-  if (
-    nodesk.detect({ provider: 'nodesk' })?.url ===
-      'https://nodesk.co/remote-jobs/index.xml'
-  ) {
-    pass('nodesk.detect() claims explicit provider config');
-  } else {
-    fail('nodesk.detect() failed');
-  }
-
-  if (nodesk.detect({ provider: 'other' }) === null) {
-    pass('nodesk.detect() ignores other provider ids');
-  } else {
-    fail('nodesk.detect() should ignore other providers');
-  }
-
-  const xml = `
-<rss>
-  <channel>
-    <item>
-      <title><![CDATA[Senior Backend Engineer at Acme]]></title>
-      <link>https://nodesk.co/remote-jobs/acme-senior-backend-engineer/</link>
-      <pubDate>Mon, 02 Jun 2025 12:00:00 GMT</pubDate>
-    </item>
-
-    <item>
-      <title>Platform Engineer at Example Corp</title>
-      <link>https://nodesk.co/remote-jobs/example-platform-engineer/</link>
-      <pubDate>not-a-date</pubDate>
-    </item>
-
-    <item>
-      <title>Bad Host at Evil Inc</title>
-      <link>https://evil.com/job</link>
-    </item>
-
-    <item>
-      <title></title>
-      <link>https://nodesk.co/remote-jobs/skip-empty-title/</link>
-    </item>
-  </channel>
-</rss>
-`;
-
-  const jobs = parseNodeskFeed(xml);
-
-  if (jobs.length === 2) {
-    pass('parseNodeskFeed keeps valid items and drops malformed ones');
-  } else {
-    fail(`expected 2 jobs, got ${jobs.length}`);
-  }
-
-  if (
-    jobs[0]?.title === 'Senior Backend Engineer' &&
-    jobs[0]?.company === 'Acme' &&
-    jobs[0]?.location === '' &&
-    jobs[0]?.url === 'https://nodesk.co/remote-jobs/acme-senior-backend-engineer/' &&
-    typeof jobs[0]?.postedAt === 'number'
-  ) {
-    pass('parseNodeskFeed maps title, company, location, url and postedAt');
-  } else {
-    fail(`unexpected first parsed job: ${JSON.stringify(jobs[0])}`);
-  }
-
-  if (
-    jobs[1]?.title === 'Platform Engineer' &&
-    jobs[1]?.company === 'Example Corp' &&
-    !('postedAt' in jobs[1])
-  ) {
-    pass('parseNodeskFeed is NaN-safe for invalid dates');
-  } else {
-    fail(`unexpected second parsed job: ${JSON.stringify(jobs[1])}`);
-  }
-
-  let fetchCalled = false;
-
-  const fetched = await nodesk.fetch({}, {
-    async fetchText(url, opts) {
-      fetchCalled = true;
-
-      if (
-        url !== 'https://nodesk.co/remote-jobs/index.xml' ||
-        opts?.redirect !== 'error'
-      ) {
-        throw new Error('unexpected fetch arguments');
-      }
-
-      return xml;
-    },
-  });
-
-  if (fetchCalled && fetched.length === 2) {
-    pass('nodesk.fetch() requests the pinned RSS feed');
-  } else {
-    fail('nodesk.fetch() did not fetch correctly');
-  }
-
-  if (fetchCalled) {
-    pass('nodesk.fetch() passes redirect:"error" to fetchText');
-  } else {
-    fail('nodesk.fetch() never called fetchText');
-  }
-
-} catch (e) {
-  fail(`nodesk provider tests crashed: ${e.message}`);
-}
-
-// ── 48. Provider — workingnomads ────────────────────────────────
-console.log('\n48. Provider — workingnomads');
-
-try {
-  const workingnomadsModule = await import(pathToFileURL(join(ROOT, 'providers/workingnomads.mjs')).href);
-  const workingnomads = workingnomadsModule.default;
-
-  if (workingnomads.id === 'workingnomads') pass('workingnomads.id is "workingnomads"');
-  else fail(`workingnomads.id is ${JSON.stringify(workingnomads.id)}`);
-
-  if (typeof workingnomads.fetch === 'function') pass('workingnomads exports a fetch() function');
-  else fail('workingnomads.fetch should be a function');
-
-  // Deterministic sample payload (top-level array) — no network. Two valid jobs
-  // plus two that must be dropped (empty title, non-absolute url). Row 0 carries
-  // surrounding whitespace on every field to verify trimming.
-  const sample = [
-    {
-      title: 'Senior AI Engineer',
-      url: 'https://www.workingnomads.com/jobs/acme-senior-ai-engineer',
-      company_name: '  Acme Corp  ',                 // surrounding space → trimmed
-      location: '  Remote (Worldwide)  ',            // surrounding space → trimmed
-    },
-    {
-      title: '  Platform Engineer  ',              // leading/trailing space → trimmed
-      url: '  https://www.workingnomads.com/jobs/beta-platform-engineer  ',
-      company_name: '',                            // empty → falls back to entry.name
-      // location omitted → ''
-    },
-    {
-      title: '',                                    // dropped: empty title
-      url: 'https://www.workingnomads.com/jobs/bad-empty-title',
-      company_name: 'Bad Co',
-    },
-    {
-      title: 'Relative URL Role',                   // dropped: non-absolute url
-      url: '/jobs/relative',
-      company_name: 'Rel Co',
-    },
-  ];
-
-  let capturedUrl = null;
-  let capturedOpts = null;
-  const fetched = await workingnomads.fetch(
-    { name: 'Working Nomads Board', provider: 'workingnomads' },
-    { fetchJson: async (url, opts) => { capturedUrl = url; capturedOpts = opts; return sample; } },
-  );
-
-  if (capturedUrl === 'https://www.workingnomads.com/api/exposed_jobs/')
-    pass('workingnomads.fetch() requests the board-wide feed URL');
-  else fail(`workingnomads.fetch() requested ${JSON.stringify(capturedUrl)}`);
-
-  if (capturedOpts && capturedOpts.redirect === 'error')
-    pass('workingnomads.fetch() passes redirect:"error" to fetchJson (SSRF guard)');
-  else fail(`workingnomads.fetch() should pass redirect:"error", got: ${JSON.stringify(capturedOpts)}`);
-
-  if (fetched.length === 2)
-    pass('workingnomads.fetch() keeps 2 valid jobs (drops empty-title + non-absolute-url rows)');
-  else fail(`workingnomads.fetch() returned ${fetched.length} jobs (expected 2)`);
-
-  // Normalized shape: exactly { title, url, company, location }.
-  if (fetched[0] && Object.keys(fetched[0]).sort().join(',') === 'company,location,title,url')
-    pass('workingnomads.fetch() returns the normalized { title, url, company, location } shape');
-  else fail(`workingnomads.fetch() row 0 keys = ${JSON.stringify(fetched[0] && Object.keys(fetched[0]))}`);
-
-  if (fetched[0]?.title === 'Senior AI Engineer'
-      && fetched[0]?.url === 'https://www.workingnomads.com/jobs/acme-senior-ai-engineer'
-      && fetched[0]?.company === 'Acme Corp'
-      && fetched[0]?.location === 'Remote (Worldwide)')
-    pass('workingnomads.fetch() maps title/url and trims company_name + location into the normalized shape');
-  else fail(`workingnomads.fetch() row 0 = ${JSON.stringify(fetched[0])}`);
-
-  if (fetched[1]?.title === 'Platform Engineer'
-      && fetched[1]?.url === 'https://www.workingnomads.com/jobs/beta-platform-engineer')
-    pass('workingnomads.fetch() trims whitespace from title and url');
-  else fail(`workingnomads.fetch() row 1 title/url = ${JSON.stringify({ title: fetched[1]?.title, url: fetched[1]?.url })}`);
-
-  if (fetched[1]?.company === 'Working Nomads Board')
-    pass('workingnomads.fetch() falls back to entry.name when company_name is empty');
-  else fail(`workingnomads.fetch() row 1 company = ${JSON.stringify(fetched[1]?.company)}`);
-
-  if (fetched[1]?.location === '')
-    pass('workingnomads.fetch() yields empty location when location is absent');
-  else fail(`workingnomads.fetch() row 1 location = ${JSON.stringify(fetched[1]?.location)}`);
-
-  // company default when both company_name and entry.name are missing → 'Working Nomads'.
-  const noName = await workingnomads.fetch(
-    {},
-    { fetchJson: async () => ([{ title: 'Role', url: 'https://www.workingnomads.com/jobs/x' }]) },
-  );
-  if (noName[0]?.company === 'Working Nomads')
-    pass('workingnomads.fetch() defaults company to "Working Nomads" when company_name and entry.name are both missing');
-  else fail(`workingnomads.fetch() default company = ${JSON.stringify(noName[0]?.company)}`);
-
-  // Empty-feed safety: an empty array yields an empty result (no crash).
-  const empty = await workingnomads.fetch({ name: 'X' }, { fetchJson: async () => ([]) });
-  if (Array.isArray(empty) && empty.length === 0) pass('workingnomads.fetch() returns [] for an empty feed');
-  else fail(`workingnomads.fetch() empty feed = ${JSON.stringify(empty)}`);
-
-  // Malformed (non-array) response → throws.
-  let badResponseThrew = false;
-  try {
-    await workingnomads.fetch(
-      { name: 'X', provider: 'workingnomads' },
-      { fetchJson: async () => ({ jobs: [] }) },
-    );
-  } catch (e) {
-    badResponseThrew = /unexpected API response/.test(e.message);
-  }
-  if (badResponseThrew) pass('workingnomads.fetch() throws on a non-array API response');
-  else fail('workingnomads.fetch() should throw when the response is not an array');
-
-} catch (e) {
-  fail(`workingnomads provider tests crashed: ${e.message}`);
-}
-
-// ── 49. Provider — 4dayweek ─────────────────────────────────────
-console.log('\n49. Provider — 4dayweek');
-
-try {
-  const fdwModule = await import(pathToFileURL(join(ROOT, 'providers/4dayweek.mjs')).href);
-  const fourdayweek = fdwModule.default;
-  const { normalize4dwJob } = fdwModule;
-
-  if (fourdayweek.id === '4dayweek') pass('4dayweek.id is "4dayweek"');
-  else fail(`4dayweek.id is ${JSON.stringify(fourdayweek.id)}`);
-
-  // normalize4dwJob — full mapping (url is BUILT from slug; feed has no url).
-  const full = normalize4dwJob(
-    { title: '  Financial Controller  ', slug: 'financial-controller-at-panzerglass-45369c18', company_name: '  PanzerGlass  ', locations: [{ city: 'Hinnerup', country: 'Denmark' }], work_arrangement: 'onsite', posted: 1782731975, is_expired: false },
-    'Fallback',
-  );
-  if (full && full.title === 'Financial Controller'
-      && full.url === 'https://4dayweek.io/job/financial-controller-at-panzerglass-45369c18'
-      && full.company === 'PanzerGlass' && full.location === 'Hinnerup, Denmark'
-      && full.postedAt === 1782731975 * 1000) {
-    pass('normalize4dwJob maps title, builds /job/<slug> url, company_name, location, posted(seconds)→ms');
-  } else {
-    fail(`normalize4dwJob full row = ${JSON.stringify(full)}`);
-  }
-
-  // work_arrangement: remote → "Remote" appended.
-  const remoteJob = normalize4dwJob({ title: 'R', slug: 'r-1', locations: [{ city: 'Berlin', country: 'Germany' }], work_arrangement: 'remote' });
-  if (remoteJob?.location === 'Berlin, Germany, Remote') pass('normalize4dwJob appends "Remote" when work_arrangement is "remote"');
-  else fail(`normalize4dwJob remote location = ${JSON.stringify(remoteJob?.location)}`);
-
-  // company fallbacks: company.name → entry name → "4 Day Week" (whitespace-only ignored).
-  const coNested = normalize4dwJob({ title: 'T', slug: 's-1', company: { name: 'Nested Co' } });
-  const coEntry = normalize4dwJob({ title: 'T', slug: 's-2' }, 'Entry Name');
-  const coDefault = normalize4dwJob({ title: 'T', slug: 's-3' });
-  const coBlank = normalize4dwJob({ title: 'T', slug: 's-4' }, '   ');
-  if (coNested?.company === 'Nested Co' && coEntry?.company === 'Entry Name'
-      && coDefault?.company === '4 Day Week' && coBlank?.company === '4 Day Week') {
-    pass('normalize4dwJob falls back company → company.name → entry name → "4 Day Week" (whitespace-only ignored)');
-  } else {
-    fail(`normalize4dwJob company fallbacks = ${JSON.stringify({ n: coNested?.company, e: coEntry?.company, d: coDefault?.company, b: coBlank?.company })}`);
-  }
-
-  // postedAt omitted when posted is absent / non-finite.
-  const noDate = normalize4dwJob({ title: 'T', slug: 's-5' });
-  const nanDate = normalize4dwJob({ title: 'T', slug: 's-6', posted: 'oops' });
-  if (noDate && !('postedAt' in noDate) && nanDate && !('postedAt' in nanDate)) {
-    pass('normalize4dwJob omits postedAt when posted is absent or non-numeric (NaN-safe)');
-  } else {
-    fail(`normalize4dwJob date handling = ${JSON.stringify({ none: noDate, nan: nanDate })}`);
-  }
-
-  // drops: expired, empty title, missing/unsafe slug, non-object.
-  const drops = [
-    normalize4dwJob({ title: 'Expired', slug: 'x-1', is_expired: true }),
-    normalize4dwJob({ title: '', slug: 'x-2' }),
-    normalize4dwJob({ title: 'No slug' }),
-    normalize4dwJob({ title: 'Unsafe slug', slug: 'a/b' }),
-    normalize4dwJob({ title: 'Spacey slug', slug: 'a b' }),
-    normalize4dwJob(null),
-  ];
-  if (drops.every(r => r === null)) {
-    pass('normalize4dwJob drops expired / empty-title / no-slug / unsafe-slug / non-object');
-  } else {
-    fail(`normalize4dwJob drops = ${JSON.stringify(drops)}`);
-  }
-
-  // fetch(): pagination by ?page=N, stop on has_more:false.
-  const mk = (i) => ({ title: `Role ${i}`, slug: `role-${i}`, company_name: `Co ${i}`, locations: [{ city: 'Lisbon', country: 'Portugal' }], posted: 1782731975 + i, is_expired: false });
-  const page1 = { jobs: Array.from({ length: 25 }, (_, i) => mk(i)), total: 50, page: 1, has_more: true };
-  const page2 = { jobs: [mk(25), mk(26), { title: '', slug: 'bad' }], total: 50, page: 2, has_more: false }; // has_more:false → stop; 1 drop
-  const requested = [];
-  const pagedFetch = async (url, opts) => {
-    requested.push({ url, redirect: opts?.redirect });
-    return Number(new URL(url).searchParams.get('page')) === 1 ? page1 : page2;
-  };
-  const paged = await fourdayweek.fetch({ name: '4 Day Week' }, { fetchJson: pagedFetch });
-
-  if (requested.length === 2
-      && requested[0].url === 'https://4dayweek.io/api/jobs?page=1'
-      && requested[1].url === 'https://4dayweek.io/api/jobs?page=2') {
-    pass('4dayweek.fetch() builds ?page=N URLs and stops when has_more is false');
-  } else {
-    fail(`4dayweek.fetch() requested = ${JSON.stringify(requested.map(r => r.url))}`);
-  }
-
-  if (requested.every(r => r.redirect === 'error')) pass('4dayweek.fetch() passes redirect:"error" on every page (SSRF guard)');
-  else fail(`4dayweek.fetch() redirect opts = ${JSON.stringify(requested.map(r => r.redirect))}`);
-
-  if (paged.length === 27) pass('4dayweek.fetch() aggregates valid jobs across pages (25 + 2, dropping the empty-title row)');
-  else fail(`4dayweek.fetch() returned ${paged.length} jobs (expected 27)`);
-
-  // max_pages cap: only the first page is requested even though has_more is true.
-  const capReq = [];
-  await fourdayweek.fetch(
-    { name: '4 Day Week', max_pages: 1 },
-    { fetchJson: async (url) => { capReq.push(url); return { jobs: Array.from({ length: 25 }, (_, i) => mk(i)), total: 999, has_more: true }; } },
-  );
-  if (capReq.length === 1 && capReq[0] === 'https://4dayweek.io/api/jobs?page=1') {
-    pass('4dayweek.fetch() honors max_pages (stops at the cap even when has_more is true)');
-  } else {
-    fail(`4dayweek.fetch() max_pages:1 requested ${JSON.stringify(capReq)}`);
-  }
-
-  // unexpected API response → throws.
-  let badThrew = false;
-  try {
-    await fourdayweek.fetch({ name: 'X' }, { fetchJson: async () => ([]) });
-  } catch (e) {
-    badThrew = /unexpected API response/.test(e.message);
-  }
-  if (badThrew) pass('4dayweek.fetch() throws on unexpected API response shape (no jobs array)');
-  else fail('4dayweek.fetch() should throw when the jobs array is absent');
-
-} catch (e) {
-  fail(`4dayweek provider tests crashed: ${e.message}`);
-}
 
 // ── Plugin engine (contract + sandbox + firewall) ────────────────
 console.log('\n49. Plugin engine (contract + sandbox + firewall)');
@@ -6046,453 +5481,6 @@ try {
   console.warn = __origWarn;
   if (__pluginTmp) { try { rmSync(__pluginTmp, { recursive: true, force: true }); } catch {} }
   if (__manifestTmp) { try { rmSync(__manifestTmp, { recursive: true, force: true }); } catch {} }
-}
-
-// -- 50. Provider - higheredjobs -----------------------------------------
-console.log('\n50. Provider - higheredjobs');
-
-try {
-  const hejModule = await import(pathToFileURL(join(ROOT, 'providers/higheredjobs.mjs')).href);
-  const higheredjobs = hejModule.default;
-  const { parseHigherEdJobsFeed } = hejModule;
-
-  if (higheredjobs.id === 'higheredjobs') pass('higheredjobs.id is "higheredjobs"');
-  else fail(`higheredjobs.id is ${JSON.stringify(higheredjobs.id)}`);
-
-  const hit = higheredjobs.detect({ name: 'HEJ', provider: 'higheredjobs', cat_id: 64 });
-  if (hit && hit.url === 'https://www.higheredjobs.com/rss/categoryFeed.cfm?catID=64') {
-    pass('higheredjobs.detect() claims explicit provider config (returns URL with catID)');
-  } else {
-    fail(`higheredjobs.detect() returned ${JSON.stringify(hit)}`);
-  }
-
-  const hitDefault = higheredjobs.detect({ name: 'HEJ', provider: 'higheredjobs' });
-  if (hitDefault && hitDefault.url === 'https://www.higheredjobs.com/rss/categoryFeed.cfm?catID=68') {
-    pass('higheredjobs.detect() with no cat_id returns default URL with catID=68');
-  } else {
-    fail(`higheredjobs.detect() default returned ${JSON.stringify(hitDefault)}`);
-  }
-
-  if (higheredjobs.detect({ name: 'Remote Board', provider: 'remoteok' }) === null) {
-    pass('higheredjobs.detect() ignores other provider ids');
-  } else {
-    fail('higheredjobs.detect() should only claim provider: higheredjobs');
-  }
-
-  const sample = `<?xml version="1.0" encoding="UTF-8"?>
-<rss version="2.0">
-  <channel>
-    <item>
-      <title><![CDATA[Director of AI Strategy]]></title>
-      <description><![CDATA[Curry College (Milton, MA)]]></description>
-      <link>https://www.higheredjobs.com/details.cfm?JobCode=17899012</link>
-      <pubDate>Thu, 13 Nov 2025 14:10:41 +0000</pubDate>
-      <guid>https://www.higheredjobs.com/details.cfm?JobCode=17899012</guid>
-    </item>
-    <item>
-      <title>Dean of Engineering &amp; Computing</title>
-      <description>State University System Office</description>
-      <link>https://www.higheredjobs.com/details.cfm?JobCode=17899044</link>
-      <pubDate>Fri, 02 Jan 2026 09:00:00 +0000</pubDate>
-      <guid>https://www.higheredjobs.com/details.cfm?JobCode=17899044</guid>
-    </item>
-    <item>
-      <title>Missing Link Role</title>
-      <description>Ghost College (Nowhere, ZZ)</description>
-      <pubDate>Fri, 02 Jan 2026 09:00:00 +0000</pubDate>
-    </item>
-    <item>
-      <title>Relative Link Role</title>
-      <description>Relative U (Somewhere, ST)</description>
-      <link>/details.cfm?JobCode=bad-relative</link>
-      <pubDate>Fri, 02 Jan 2026 09:00:00 +0000</pubDate>
-    </item>
-    <item>
-      <title>Off Host Role</title>
-      <description>Off Host Inst (Elsewhere, ST)</description>
-      <link>https://example.com/details.cfm?JobCode=off-host</link>
-      <pubDate>Fri, 02 Jan 2026 09:00:00 +0000</pubDate>
-    </item>
-  </channel>
-</rss>`;
-  const jobs = parseHigherEdJobsFeed(sample, 'HEJ Board');
-
-  if (jobs.length === 2) pass('parseHigherEdJobsFeed keeps 2 items (drops missing/relative/off-host links)');
-  else fail(`parseHigherEdJobsFeed returned ${jobs.length} jobs (expected 2)`);
-
-  if (jobs.every(({ url }) => {
-    try {
-      const parsed = new URL(url);
-      return parsed.protocol === 'https:' && parsed.hostname === 'www.higheredjobs.com';
-    } catch {
-      return false;
-    }
-  })) pass('parseHigherEdJobsFeed only emits HTTPS URLs pinned to www.higheredjobs.com');
-  else fail('parseHigherEdJobsFeed emitted an off-host or non-HTTPS URL');
-
-  if (jobs[0]?.title === 'Director of AI Strategy' && jobs[0]?.company === 'Curry College' && jobs[0]?.location === 'Milton, MA') {
-    pass('parseHigherEdJobsFeed parses title + "Institution (City, ST)" description -> company/title/location');
-  } else {
-    fail(`row 0 = ${JSON.stringify(jobs[0])}`);
-  }
-
-  if (jobs[0]?.url === 'https://www.higheredjobs.com/details.cfm?JobCode=17899012') {
-    pass('parseHigherEdJobsFeed maps <link> to url');
-  } else {
-    fail(`row 0 url = ${JSON.stringify(jobs[0]?.url)}`);
-  }
-
-  if (jobs[0]?.postedAt === Date.parse('Thu, 13 Nov 2025 14:10:41 +0000')) {
-    pass('parseHigherEdJobsFeed parses pubDate -> postedAt');
-  } else {
-    fail(`row 0 postedAt = ${JSON.stringify(jobs[0]?.postedAt)}`);
-  }
-
-  if (jobs[1]?.company === 'State University System Office' && jobs[1]?.location === '' && jobs[1]?.title === 'Dean of Engineering & Computing') {
-    pass('parseHigherEdJobsFeed falls back to whole description as company when no parens (empty location)');
-  } else {
-    fail(`row 1 = ${JSON.stringify(jobs[1])}`);
-  }
-
-  if (parseHigherEdJobsFeed('', 'X').length === 0 && parseHigherEdJobsFeed(null, 'X').length === 0) {
-    pass('parseHigherEdJobsFeed empty / non-string feed -> empty result (no crash)');
-  } else {
-    fail('parseHigherEdJobsFeed empty / non-string feed should yield empty result');
-  }
-
-  let capturedUrl = null;
-  let capturedOpts = null;
-  const fetched = await higheredjobs.fetch(
-    { name: 'HEJ Board', provider: 'higheredjobs', cat_id: 64 },
-    { fetchText: async (url, opts) => { capturedUrl = url; capturedOpts = opts; return sample; } },
-  );
-
-  if (capturedUrl === 'https://www.higheredjobs.com/rss/categoryFeed.cfm?catID=64') {
-    pass('higheredjobs.fetch() requests the feed URL for cat_id 64');
-  } else {
-    fail(`higheredjobs.fetch() requested ${JSON.stringify(capturedUrl)}`);
-  }
-
-  if (capturedOpts && capturedOpts.redirect === 'error') {
-    pass('higheredjobs.fetch() passes redirect:"error" to fetchText');
-  } else {
-    fail(`higheredjobs.fetch() should pass redirect:"error", got: ${JSON.stringify(capturedOpts)}`);
-  }
-
-  if (fetched[0]?.company === 'Curry College' && fetched[0]?.title === 'Director of AI Strategy' && fetched[0]?.location === 'Milton, MA') {
-    pass('provider: higheredjobs config returns normalized jobs');
-  } else {
-    fail(`higheredjobs.fetch() normalized row = ${JSON.stringify(fetched[0])}`);
-  }
-} catch (e) {
-  fail(`higheredjobs provider tests crashed: ${e.message}`);
-}
-
-// ── 51. PROVIDERS — JibeApply ────────────────────────────────────
-
-console.log('\n51. Provider — jibeapply');
-
-try {
-  const jibeapply = (await import(pathToFileURL(join(ROOT, 'providers/jibeapply.mjs')).href)).default;
-  const { parseJibeapplyResponse } = await import(pathToFileURL(join(ROOT, 'providers/jibeapply.mjs')).href);
-
-  if (jibeapply.id === 'jibeapply') pass('jibeapply.id is "jibeapply"');
-  else fail(`jibeapply.id is ${JSON.stringify(jibeapply.id)}`);
-
-  // detect() — /jobs path → /api/jobs
-  const hit = jibeapply.detect({ name: 'Acme', careers_url: 'https://acme.jibeapply.com/jobs?location=Germany' });
-  if (hit && hit.url === 'https://acme.jibeapply.com/api/jobs?location=Germany') {
-    pass('jibeapply.detect() rewrites /jobs → /api/jobs');
-  } else {
-    fail(`jibeapply.detect() returned ${JSON.stringify(hit)}`);
-  }
-
-  // detect() — /api/jobs already present (idempotent)
-  const hitApi = jibeapply.detect({ name: 'X', careers_url: 'https://acme.jibeapply.com/api/jobs' });
-  if (hitApi && hitApi.url === 'https://acme.jibeapply.com/api/jobs') {
-    pass('jibeapply.detect() leaves already-correct /api/jobs URL unchanged');
-  } else {
-    fail(`jibeapply.detect(api) returned ${JSON.stringify(hitApi)}`);
-  }
-
-  // detect() — null cases
-  if (jibeapply.detect({ name: 'X', careers_url: 'https://example.com/jobs' }) === null) {
-    pass('jibeapply.detect() returns null for non-jibeapply URL');
-  } else {
-    fail('jibeapply.detect() should return null for non-jibeapply URL');
-  }
-
-  // Path-spoofed: jibeapply.com in path, not host
-  if (jibeapply.detect({ name: 'Spoof', careers_url: 'https://evil.example/acme.jibeapply.com/jobs' }) === null) {
-    pass('jibeapply.detect() rejects path-spoofed URL');
-  } else {
-    fail('jibeapply.detect() must NOT detect path-spoofed URLs');
-  }
-
-  // Non-string careers_url
-  if (jibeapply.detect({ name: 'X', careers_url: 42 }) === null) {
-    pass('jibeapply.detect() returns null for non-string careers_url');
-  } else {
-    fail('jibeapply.detect() should return null for non-string careers_url');
-  }
-
-  // HTTP (non-HTTPS) must not be detected
-  if (jibeapply.detect({ name: 'X', careers_url: 'http://acme.jibeapply.com/jobs' }) === null) {
-    pass('jibeapply.detect() rejects HTTP (non-HTTPS) URL');
-  } else {
-    fail('jibeapply.detect() should reject HTTP URLs');
-  }
-
-  // parseJibeapplyResponse — normalization
-  const entry = { name: 'Acme', careers_url: 'https://acme.jibeapply.com/jobs' };
-  const sampleJson = {
-    jobs: [
-      { title: 'Senior QA', slug: 'senior-qa-berlin', city: 'Berlin', country: 'Germany', hiring_organization: 'Acme GmbH' },
-      { title: 'Backend Dev', req_id: 'REQ-123', full_location: 'Remote, Germany' },
-      { data: { title: 'Wrapped Job', slug: 'wrapped-job', city: 'Munich', country: 'Germany' } },
-      { title: '', slug: 'no-title' },          // missing title — skip
-      { title: 'No Slug' },                      // missing slug/req_id — skip
-    ],
-  };
-  const parsedJibe = parseJibeapplyResponse(sampleJson, entry);
-
-  if (parsedJibe.length === 3) pass('parseJibeapplyResponse extracts 3 valid jobs');
-  else fail(`parseJibeapplyResponse returned ${parsedJibe.length} jobs, expected 3`);
-
-  if (parsedJibe[0].url === 'https://acme.jibeapply.com/jobs/senior-qa-berlin') {
-    pass('parseJibeapplyResponse builds URL from origin + /jobs/ + slug');
-  } else {
-    fail(`row 0 url = ${JSON.stringify(parsedJibe[0].url)}`);
-  }
-
-  if (parsedJibe[0].location === 'Berlin, Germany') {
-    pass('parseJibeapplyResponse builds location from city/country');
-  } else {
-    fail(`row 0 location = ${JSON.stringify(parsedJibe[0].location)}`);
-  }
-
-  if (parsedJibe[0].company === 'Acme GmbH') {
-    pass('parseJibeapplyResponse uses hiring_organization when present');
-  } else {
-    fail(`row 0 company = ${JSON.stringify(parsedJibe[0].company)}`);
-  }
-
-  if (parsedJibe[1].url.includes('REQ-123') && parsedJibe[1].location === 'Remote, Germany') {
-    pass('parseJibeapplyResponse uses req_id and full_location');
-  } else {
-    fail(`row 1 = ${JSON.stringify(parsedJibe[1])}`);
-  }
-
-  if (parsedJibe[2].title === 'Wrapped Job') {
-    pass('parseJibeapplyResponse unwraps item.data');
-  } else {
-    fail(`row 2 title = ${JSON.stringify(parsedJibe[2].title)}`);
-  }
-
-  // parseJibeapplyResponse — falls back to entry.name when hiring_organization missing
-  const noOrg = parseJibeapplyResponse({ jobs: [{ title: 'Dev', slug: 'dev', city: 'Berlin', country: 'Germany' }] }, entry);
-  if (noOrg[0].company === 'Acme') {
-    pass('parseJibeapplyResponse falls back to entry.name when hiring_organization missing');
-  } else {
-    fail(`fallback company = ${JSON.stringify(noOrg[0].company)}`);
-  }
-
-  // parseJibeapplyResponse — empty input
-  if (parseJibeapplyResponse({}, entry).length === 0) pass('parseJibeapplyResponse({}) → empty result');
-  else fail('parseJibeapplyResponse({}) should be empty');
-
-  // parseJibeapplyResponse — falls back to entry.api's origin when careers_url
-  // can't be parsed, so job URLs stay absolute instead of degrading to "/jobs/<slug>"
-  const malformedCareersEntry = { name: 'Widget Co', careers_url: 'jobs.widgetco.com', api: 'https://jobs.widgetco.com/api/jobs' };
-  const apiOriginFallback = parseJibeapplyResponse({ jobs: [{ title: 'Dev', slug: 'dev-1' }] }, malformedCareersEntry);
-  if (apiOriginFallback[0]?.url === 'https://jobs.widgetco.com/jobs/dev-1') {
-    pass('parseJibeapplyResponse falls back to entry.api origin for a malformed careers_url');
-  } else {
-    fail(`parseJibeapplyResponse api-origin fallback: url = ${JSON.stringify(apiOriginFallback[0]?.url)}`);
-  }
-
-  // parseJibeapplyResponse — null/undefined entries in jobs must be skipped, not crash
-  const sparseJson = { jobs: [null, undefined, { title: 'Real Job', slug: 'real-job' }] };
-  try {
-    const parsedSparse = parseJibeapplyResponse(sparseJson, entry);
-    if (parsedSparse.length === 1 && parsedSparse[0].title === 'Real Job') {
-      pass('parseJibeapplyResponse skips null/undefined entries without crashing');
-    } else {
-      fail(`parseJibeapplyResponse sparse result = ${JSON.stringify(parsedSparse)}`);
-    }
-  } catch (e3) {
-    fail(`parseJibeapplyResponse should not throw on null/undefined entries: ${e3.message}`);
-  }
-
-  // fetch() pagination — 2 pages
-  let pageRequests = 0;
-  const fetchedJibe = await jibeapply.fetch(entry, {
-    transport: 'http',
-    fetchText: async () => { throw new Error('fetchText not expected'); },
-    fetchJson: async (url) => {
-      pageRequests++;
-      const u = new URL(url);
-      const page = parseInt(u.searchParams.get('page') || '1', 10);
-      if (page === 1) {
-        return { totalCount: 15, count: 10, jobs: Array.from({ length: 10 }, (_, i) => ({ title: `Job ${i}`, slug: `job-${i}` })) };
-      }
-      return { totalCount: 15, count: 10, jobs: Array.from({ length: 5 }, (_, i) => ({ title: `Job p2-${i}`, slug: `job-p2-${i}` })) };
-    },
-  });
-  if (pageRequests === 2 && fetchedJibe.length === 15) {
-    pass('jibeapply.fetch() paginates across 2 pages (10+5=15)');
-  } else {
-    fail(`jibeapply fetch pagination: requests=${pageRequests}, total=${fetchedJibe.length} (expected 2/15)`);
-  }
-
-  // fetch() pagination cap — an inflated totalCount must not trigger unbounded
-  // concurrent requests (MAX_PAGES = 50 in providers/jibeapply.mjs), and
-  // hitting the cap must be visible (console.error), not silent.
-  let hugeRequests = 0;
-  const jibeCapWarnings = [];
-  const originalConsoleError = console.error;
-  console.error = (msg) => jibeCapWarnings.push(msg);
-  let fetchedHuge;
-  try {
-    fetchedHuge = await jibeapply.fetch(entry, {
-      transport: 'http',
-      fetchText: async () => { throw new Error('fetchText not expected'); },
-      fetchJson: async () => {
-        hugeRequests++;
-        return { totalCount: 1_000_000, count: 10, jobs: Array.from({ length: 10 }, (_, i) => ({ title: `Job ${i}`, slug: `job-${i}` })) };
-      },
-    });
-  } finally {
-    console.error = originalConsoleError;
-  }
-  if (hugeRequests === 50 && fetchedHuge.length === 500) {
-    pass('jibeapply.fetch() caps pagination at MAX_PAGES despite an inflated totalCount');
-  } else {
-    fail(`jibeapply fetch pagination cap: requests=${hugeRequests}, total=${fetchedHuge.length} (expected 50/500)`);
-  }
-  if (jibeCapWarnings.some(w => /has more postings than max_pages allows/.test(w))) {
-    pass('jibeapply.fetch() warns (console.error) when the cap truncates real results');
-  } else {
-    fail(`jibeapply fetch cap: expected a truncation warning, got ${JSON.stringify(jibeCapWarnings)}`);
-  }
-
-  // fetch() pagination cap — entry.max_pages raises the cap for a genuinely
-  // large tenant (e.g. a Workday-scale Deutsche Bank equivalent on JibeApply)
-  let overriddenRequests = 0;
-  const bigEntry = { name: 'BigCo', careers_url: 'https://bigco.jibeapply.com/jobs', max_pages: 80 };
-  const fetchedOverridden = await jibeapply.fetch(bigEntry, {
-    transport: 'http',
-    fetchText: async () => { throw new Error('fetchText not expected'); },
-    fetchJson: async () => {
-      overriddenRequests++;
-      return { totalCount: 1_000_000, count: 10, jobs: Array.from({ length: 10 }, (_, i) => ({ title: `Job ${i}`, slug: `job-${i}` })) };
-    },
-  });
-  if (overriddenRequests === 80 && fetchedOverridden.length === 800) {
-    pass('jibeapply.fetch() honors entry.max_pages to raise the cap above the default');
-  } else {
-    fail(`jibeapply fetch max_pages override: requests=${overriddenRequests}, total=${fetchedOverridden.length} (expected 80/800)`);
-  }
-
-  // entry.max_pages is itself capped (MAX_PAGES_CAP = 500) — an absurd override
-  // can't turn this into an unbounded scan either
-  let cappedOverrideRequests = 0;
-  const absurdEntry = { name: 'AbsurdCo', careers_url: 'https://absurdco.jibeapply.com/jobs', max_pages: 100_000 };
-  const fetchedAbsurd = await jibeapply.fetch(absurdEntry, {
-    transport: 'http',
-    fetchText: async () => { throw new Error('fetchText not expected'); },
-    fetchJson: async () => {
-      cappedOverrideRequests++;
-      return { totalCount: 10_000_000, count: 10, jobs: Array.from({ length: 10 }, (_, i) => ({ title: `Job ${i}`, slug: `job-${i}` })) };
-    },
-  });
-  if (cappedOverrideRequests === 500 && fetchedAbsurd.length === 5000) {
-    pass('jibeapply.fetch() caps an absurd entry.max_pages at MAX_PAGES_CAP');
-  } else {
-    fail(`jibeapply fetch max_pages hard cap: requests=${cappedOverrideRequests}, total=${fetchedAbsurd.length} (expected 500/5000)`);
-  }
-
-  // fetch() pagination — a failure on a later page returns the jobs gathered
-  // so far instead of discarding everything (sequential, not Promise.all),
-  // and the failure itself is visible (console.error), not silent.
-  let flakyRequests = 0;
-  const jibeFlakyWarnings = [];
-  console.error = (msg) => jibeFlakyWarnings.push(msg);
-  let fetchedFlaky;
-  try {
-    fetchedFlaky = await jibeapply.fetch(entry, {
-      transport: 'http',
-      fetchText: async () => { throw new Error('fetchText not expected'); },
-      fetchJson: async (url) => {
-        flakyRequests++;
-        const u = new URL(url);
-        const page = parseInt(u.searchParams.get('page') || '1', 10);
-        if (page === 3) throw new Error('HTTP 503');
-        return { totalCount: 40, count: 10, jobs: Array.from({ length: 10 }, (_, i) => ({ title: `Job p${page}-${i}`, slug: `job-p${page}-${i}` })) };
-      },
-    });
-  } finally {
-    console.error = originalConsoleError;
-  }
-  if (flakyRequests === 3 && fetchedFlaky.length === 20) {
-    pass('jibeapply.fetch() returns partial results when a later page fails');
-  } else {
-    fail(`jibeapply fetch partial failure: requests=${flakyRequests}, total=${fetchedFlaky.length} (expected 3/20)`);
-  }
-  if (jibeFlakyWarnings.some(w => /page \d+ fetch failed/.test(w))) {
-    pass('jibeapply.fetch() warns (console.error) when a page fetch fails mid-pagination');
-  } else {
-    fail(`jibeapply fetch page failure: expected a fetch-failed warning, got ${JSON.stringify(jibeFlakyWarnings)}`);
-  }
-
-  // fetch() with explicit entry.api (non-jibeapply.com host)
-  let explicitApiUrl = null;
-  const brandedEntry = { name: 'Widget Co', careers_url: 'https://jobs.widgetco.com/jobs', api: 'https://jobs.widgetco.com/api/jobs?internal=false' };
-  await jibeapply.fetch(brandedEntry, {
-    transport: 'http',
-    fetchText: async () => { throw new Error('fetchText not expected'); },
-    fetchJson: async (url) => { explicitApiUrl = url; return { totalCount: 3, count: 3, jobs: [{ title: 'Dev', slug: 'dev-1' }] }; },
-  });
-  if (explicitApiUrl && explicitApiUrl.startsWith('https://jobs.widgetco.com/api/jobs')) {
-    pass('jibeapply.fetch() uses entry.api for non-jibeapply.com hosts');
-  } else {
-    fail(`jibeapply.fetch() with entry.api called url=${JSON.stringify(explicitApiUrl)}`);
-  }
-
-  // fetch() with entry.api — iCIMS-hosted JibeApply pattern: count === totalCount but jobs.length < count
-  let brandedRequests = 0;
-  const fetchedBranded = await jibeapply.fetch(brandedEntry, {
-    transport: 'http',
-    fetchText: async () => { throw new Error('fetchText not expected'); },
-    fetchJson: async (url) => {
-      brandedRequests++;
-      const u = new URL(url);
-      const page = parseInt(u.searchParams.get('page') || '1', 10);
-      // count === totalCount (iCIMS-hosted pattern), jobs only has page-worth of items
-      if (page === 1) return { totalCount: 15, count: 15, jobs: Array.from({ length: 10 }, (_, i) => ({ title: `J${i}`, slug: `j-${i}` })) };
-      return { totalCount: 15, count: 15, jobs: Array.from({ length: 5 }, (_, i) => ({ title: `J2-${i}`, slug: `j2-${i}` })) };
-    },
-  });
-  if (brandedRequests === 2 && fetchedBranded.length === 15) {
-    pass('jibeapply.fetch() paginates when count===totalCount but jobs.length < count');
-  } else {
-    fail(`jibeapply fetch iCIMS pattern: requests=${brandedRequests}, total=${fetchedBranded.length} (expected 2/15)`);
-  }
-
-  // fetch() throws when both entry.api is HTTP and careers_url is non-jibeapply.com
-  // (no valid API URL can be derived from either source)
-  try {
-    await jibeapply.fetch({ name: 'X', careers_url: 'https://jobs.example.com/jobs', api: 'http://evil.com/api/jobs' }, {
-      fetchText: async () => '', fetchJson: async () => { throw new Error('should not reach'); },
-    });
-    fail('jibeapply.fetch() should throw when no valid API URL can be derived');
-  } catch (e2) {
-    if (/cannot derive API URL/i.test(e2.message)) pass('jibeapply.fetch() throws when HTTP entry.api and non-jibeapply careers_url');
-    else fail(`jibeapply.fetch() wrong error: ${e2.message}`);
-  }
-
-} catch (e) {
-  fail(`jibeapply provider tests crashed: ${e.message}`);
 }
 
 // ── 52. INTERVIEW SESSION PRODUCER (#956 / #1242 contract) ──────
